refactored project structure to match go conventions
