+++ internal/courses/requirements.go
@@ -1,4 +1,4 @@
-package Courses
+package courses
 
 type CourseRequirementRules []CourseRequirementRule
