+++ internal/degrees/degrees_test.go
@@ -1,0 +1,14 @@
+package degrees
+
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestRegularDegree_Plan(t *testing.T) {
+	assert.Equal(t, 0, 0)
+}
+
+func TestRegularDegree_IsCompleted(t *testing.T) {
+	assert.Equal(t, 0, 0)
+}
