+++ internal/utils/mapUtils.go
@@ -1,4 +1,4 @@
-package Utils
+package utils
 
 func InMap(val interface{}, mp map[interface{}]interface{}) bool {
 	_, ok := mp[val]
